Calculator class refactored following code review:

+++ src/main/java/Calculator.java
@@ -10,7 +10,7 @@ public class Calculator {
     public int multiply(int num1, int num2){
         return num1*num2;
     }
-    public double divide(int num1, int num2){
+    public double divide(double num1, double num2){
         return num1/num2;
     }
 }
